flask_website/requirements.txt
flask[async]==3.0.2
flask-cors==4.0.0
pandas==2.2.1
numpy==1.26.4
sentence-transformers==2.5.1
google-generativeai==0.3.2
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
python-dotenv>=1.0.0
pymongo==4.13.0
fastapi==0.109.2
uvicorn==0.27.1
neo4j>=5.25.0
kafka-python==2.0.2
kafka-python-ng>=2.0.3
google-cloud-storage>=2.14.0
APScheduler>=3.10.4